# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Fashion Scraper Argentina API
  version: "1.0.0"
  description: >-
    The mechanical HTTP surface of the backend: path, method, auth requirement
    and status codes. This file is hand-maintained — there is no command that
    regenerates it — and is cross-checked
    against `ApiRoutePolicy` by `OpenApiRouteCoverageTest`.


    Envelope: every JSON success is `{ "data": T }` (lists add `"page":
    { number, size, total, totalPages }`, 0-based) and every error is
    `{ "error": { code, message, details? } }`. Not wrapped: `GET /api/csv`
    and `GET /api/openapi.yaml`. A body-less `204` stays body-less. The
    per-operation `data` shape is described in each response's description.


    Guard limit — read this before trusting a green build for more than it
    proves: this contract and its test guarantee path, method and access-level
    parity ONLY. They never assert response-shape accuracy — only a human
    review of the shapes below can. Rationale (the "why" behind auth
    semantics, per-owner scoping, and specific status codes) lives in
    `docs/API_REFERENCE.md`, not here.
servers:
  - url: http://localhost:3000
    description: Default local backend (portable/POSIX install)
security:
  - bearerAuth: []
...
components:
  schemas:
...
    ApiResponse:
      type: object
      description: Success envelope. `data` holds the operation's payload.
      required: [data]
      properties:
        data: {}
...
    ApiErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable code (`no_autenticado`, `sin_permiso`, `solicitud_invalida`, `no_encontrado`, ...).
        message: { type: string }
        details:
          type: object
          additionalProperties: true
          description: Optional structured context, e.g. `{ actual }` for `conflicto_stale`.
    ApiError:
      type: object
      description: Error envelope, used by every non-2xx JSON response.
      required: [error]
      properties:
        error: { $ref: "#/components/schemas/ApiErrorBody" }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        The default scheme for every gated route. `x-access: AUTHENTICATED`
        accepts any valid subject (ADMIN or VIEWER); `x-access: ADMIN` further
        requires the ADMIN role, re-read from the database on every request —
        never from a JWT claim.
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh
      description: >-
        `HttpOnly; Secure; SameSite=Strict; Path=/api/auth/refresh`. Used
        instead of `bearerAuth` on the two `/api/auth/refresh` operations.
        Also requires the `X-Refresh-CSRF` header (see the `RefreshCsrf`
        parameter) except on the documented cold-boot admission path — see
        `docs/API_REFERENCE.md`.
  parameters:
...
    RefreshCsrf:
      name: X-Refresh-CSRF
      in: header
      required: false
      schema: { type: string }
      description: >-
        Required except on the documented cold-boot admission path (no
        header at all, plus `Origin` + `Sec-Fetch-Site` checks) — see
        `docs/API_REFERENCE.md`.
  responses:
    Unauthorized:
      description: Missing, expired or invalid access token (`no_autenticado`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
    TooManyRequests:
      description: Rate limit exceeded.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Authenticate by username; issues a 15-minute access token.
      x-access: PERMIT
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        "200":
          description: "`data`: `{ accessToken, tokenType, expiresIn }`"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          description: >-
            Same status and body for wrong credential, unknown user,
            deactivated account, or malformed body — never distinguished.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate the refresh-token family; issues a successor access token.
      x-access: PERMIT
      security:
        - refreshCookie: []
      parameters:
        - $ref: "#/components/parameters/RefreshCsrf"
      responses:
        "200":
          description: "`data`: `{ accessToken, tokenType, expiresIn, csrfNonce }` + `Set-Cookie`"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          description: "`refresh_invalido` or `sesion_invalidada`"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: "`csrf_invalido` — missing/mismatched nonce; the presented token is left intact."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      tags: [Auth]
      summary: Logout — revokes the entire refresh-token family.
      x-access: PERMIT
      security:
        - refreshCookie: []
      parameters:
        - $ref: "#/components/parameters/RefreshCsrf"
      responses:
        "200":
          description: Cookie cleared unconditionally, even for an unknown token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "403":
          description: "`csrf_invalido`"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/auth/me:
    get:
      tags: [Auth]
      summary: The authenticated subject.
      x-access: AUTHENTICATED
      responses:
        "200":
          description: "`data`: `{ username, roles: string[] }`"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/password-reset/request:
    post:
      tags: [Auth]
      summary: "\"Forgot password\" — always 202, same body, same speed."
      x-access: PERMIT
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
      responses:
        "202":
          description: Identical body for a real, unknown, malformed or service-account address.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /api/auth/password-reset/confirm:
    post:
      tags: [Auth]
      summary: Consume a reset token and set a new password.
      x-access: PERMIT
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token: { type: string }
                password: { type: string }
      responses:
        "200":
          description: Password changed; every session for the user is revoked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          description: "`reseteo_invalido` — unknown, used, expired token, or password under 8 chars."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Usuarios ────────────────────────────────────────────────────────────
...